Fix erroneous deprecation warnings caused by including onedim.h
