+++ include/cantera/oneD/Boundary1D.h
@@ -97,6 +97,21 @@ protected:
 };
 
 
+/*!
+ * Renamed base class for boundaries between one-dimensional spatial domains.
+ * @deprecated To be removed after Cantera 2.5.
+ */
+class Bdry1D : public Boundary1D
+{
+public:
+    Bdry1D() : Boundary1D() {
+        warn_deprecated("Bdry1D::Bdry1D()",
+                        "To be removed after Cantera 2.5. "
+                        "Class renamed to Boundary1D.");
+    }
+};
+
+
 /**
  * An inlet.
  * @ingroup onedim
