+++ include/cantera/onedim.h
@@ -14,7 +14,6 @@
 #include "oneD/Sim1D.h"
 #include "oneD/Domain1D.h"
 #include "oneD/Boundary1D.h"
-#include "oneD/Inlet1D.h"
 #include "oneD/StFlow.h"
 
 #endif
